All tests possitive for IP addition
