+++ app/Http/Controllers/PageGenerationController.php
@@ -11,7 +11,6 @@ use Illuminate\Support\Facades\DB;
 
 class PageGenerationController extends Controller
 {
-
 	public static function CreateUserFile(string $name){
 		if(preg_match('/(\\.|\\/|;)/', $name))
 			return response(json_encode(["warn"=>"Name has invalid characters"]), 422)->header('Content-Type', 'text/plain');
@@ -23,7 +22,7 @@ class PageGenerationController extends Controller
 		return $fname;
 	}
 
-	public static function getLimitedInfo(){	
+	public static function getLimitedInfo(){
 		$data = (array)PageGenerationController::GetLimitedEntries();
 		$data = array_reverse(array_pop($data));
 		return json_encode($data);
@@ -50,42 +49,39 @@ class PageGenerationController extends Controller
 	// banned users will not show up in rotation
 	public static function GetRandomAdEntry(){
 		try{
-                        $name = auth()->setToken($request->cookie('freeadstoken'))->user()->name;
+                        $name = auth()->setToken($_COOKIE['freeadstoken'])->user()->name;
                 }
                 catch(\Exception $e){
-                        $name = null;           
+                        $name = "";           
 		}
-                return DB::table('ads')->
-			leftJoin('bans', 'ads.fk_name', '=', 'bans.fk_name')
-			->when(!PageGenerationController::checkBanned($name), function($q){
-				return $q->whereNull('bans.hardban')
-					->where('ip','=', $_SERVER["HTTP_X_REAL_IP"]);	
-			})		
+                return DB::table('ads')
+			->leftJoin('bans', 'ads.fk_name', '=', 'bans.fk_name')
+			->whereNull('bans.hardban')
+			->orWhere('ip','=', $_SERVER["HTTP_X_REAL_IP"])	
 			->select("ads.fk_name", "uri", "url")
-			->get();
+			->orderBy('ads.created_at', 'ASC')->inRandomOrder()->first();
 	
 		}
 
 	public static function GetLimitedEntries($name = null){
 		try{
 			if(!$name)
-                        	$name = auth()->setToken($request->cookie('freeadstoken'))->user()->name;
+                        	$name = auth()->setToken($_COOKIE['freeadstoken'])->user()->name;
                 }
                 catch(\Exception $e){
-                        $name = null;           
+                        $name = "";           
 		}
-                return DB::table('ads')->
-			leftJoin('bans', 'ads.fk_name', '=', 'bans.fk_name')
-			->when(!PageGenerationController::checkBanned($name), function($q){
+		return DB::table('ads')
+			->leftJoin('bans', 'ads.fk_name', '=', 'bans.fk_name')
+			->when($name == "" || !PageGenerationController::checkBanned($name),function($q){
 				return $q->whereNull('bans.hardban')
-					->where('ip','=', $_SERVER["HTTP_X_REAL_IP"]);	
-			})		
+					->orWhere('ip','=', $_SERVER["HTTP_X_REAL_IP"]);
+			})
 			->select("ads.fk_name", "uri", "url")
-			->get();
-	}
+			->orderBy('ads.created_at', 'ASC')->get();
+                }
 
 	public static function checkBanned($name){
 		return  DB::table('bans')->where('fk_name', '=', $name)->count() > 0;
-
 	}
 }
